[clang-tidy] Fixes for LLVM 21 (#102)
* Adjust .clang-tidy for LLVM 21

* Disable misc-predictable-rand

We want a predictable PRNG.

* Disable llvm-header-guard, portability-avoid-pragma-once:

We allow use of #pragma once.

* Disable bugprone-macro-parentheses

Some macro parameters cannot be parenthesized without producing invalid
programs. For instance, macros taking a TYPE parameter that use it to
declare a variable of type TYPE.

* Disable clang-analyzer-optin.performance.Padding

This checker warns against suboptimal padding in structs. Since
non-optimal structs are currently only present in test harnesses, and
that their performance or footprint is not critical, we disable the
check.

* Remove C++/ObjC and aliased checkers

cert-* checkers are aliases of other checkers already.
google-* checkers are either aliases, C++/ObjC checkers, or
googletest-specific checkers.
hicpp-* checkers are deprecated C++ checkers and getting removed in LLVM
23.

* Fix *-avoid-const-params-in-decls

* Fix readability-redundant-parentheses

* Fix bugprone-narrowing-conversion

* Fix readability-inconsistent-declaration-parameter-name

* Remove examples temporarily.

* Silence bugprone-branch-clone false positives

clang-tidy doesn't parse inline assembly templates, so when the output
and input operands are the same in each side of an if-else block, the
warning is raised even if the templates are indeed different. This
happens in Xsfmm gemm kernels.

Rather than disabling the warning, which may be useful, silence it where
it occured.

+++ src/gemm/xsfmm/gemm_a1b01_bf16c_bf16_f32_xsfmm32a16f.c
@@ -146,9 +146,9 @@ SKL_FUNC_PRIVATE void skl_gemm_1tm2tn_2tm1tn_a1b01_bf16c_bf16_f32_xsfmm32a16f(
 
   const size_t tm0 = bta ? tn : tm;
   const size_t tn0 = bta ? tm : tn;
-  const size_t tss_pattern = bta ? (size_t)(1) << kShiftPattern : 0;
+  const size_t tss_pattern = bta ? 1 << kShiftPattern : 0;
   const size_t mt0 = 0 | tss_pattern;
-  const size_t mt4 = (size_t)(4) << kShiftTile | tss_pattern;
+  const size_t mt4 = 4 << kShiftTile | tss_pattern;
 
   float *c0 = c;
   float *c1 = c0 + tsc1;
@@ -355,10 +355,10 @@ SKL_FUNC_PRIVATE void skl_gemm_1tm3tn_3tm1tn_a1b01_bf16c_bf16_f32_xsfmm32a16f(
 
   const size_t tm0 = bta ? tn : tm;
   const size_t tn0 = bta ? tm : tn;
-  const size_t tss_pattern = bta ? (size_t)(1) << kShiftPattern : 0;
+  const size_t tss_pattern = bta ? 1 << kShiftPattern : 0;
   const size_t mt0 = 0 | tss_pattern;
-  const size_t mt4 = (size_t)(4) << kShiftTile | tss_pattern;
-  const size_t mt8 = (size_t)(8) << kShiftTile | tss_pattern;
+  const size_t mt4 = 4 << kShiftTile | tss_pattern;
+  const size_t mt8 = 8 << kShiftTile | tss_pattern;
 
   float *c0 = c;
   float *c1 = c0 + tsc1;
@@ -605,11 +605,11 @@ SKL_FUNC_PRIVATE void skl_gemm_1tm4tn_4tm1tn_a1b01_bf16c_bf16_f32_xsfmm32a16f(
 
   const size_t tm0 = bta ? tn : tm;
   const size_t tn0 = bta ? tm : tn;
-  const size_t tss_pattern = bta ? (size_t)(1) << kShiftPattern : 0;
+  const size_t tss_pattern = bta ? 1 << kShiftPattern : 0;
   const size_t mt0 = 0 | tss_pattern;
-  const size_t mt4 = (size_t)(4) << kShiftTile | tss_pattern;
-  const size_t mt8 = (size_t)(8) << kShiftTile | tss_pattern;
-  const size_t mt12 = (size_t)(12) << kShiftTile | tss_pattern;
+  const size_t mt4 = 4 << kShiftTile | tss_pattern;
+  const size_t mt8 = 8 << kShiftTile | tss_pattern;
+  const size_t mt12 = 12 << kShiftTile | tss_pattern;
 
   float *c0 = c;
   float *c1 = c0 + tsc1;
@@ -967,9 +967,9 @@ SKL_FUNC_PRIVATE void skl_gemm_2tn2tn_a1b01_bf16pc_bf16cp_f32rcp_xsfmm32a16f(
   const size_t kRowInc = 1;
 
   const size_t mt0 = 0;
-  const size_t mt4 = (size_t)(4) << kShiftTile;
-  const size_t mt8 = (size_t)(8) << kShiftTile;
-  const size_t mt12 = (size_t)(12) << kShiftTile;
+  const size_t mt4 = 4 << kShiftTile;
+  const size_t mt8 = 8 << kShiftTile;
+  const size_t mt12 = 12 << kShiftTile;
 
   float *c00 = c;
   float *c01 = c00 + csc1;
